move callController to ts-rest-adapter.ts
