+++ backend/src/api/routes/index.ts
@@ -178,74 +178,3 @@ function applyApiRoutes(app: Application): void {
     app.use(apiRoute, router);
   });
 }
-
-export function callController<
-  TRoute extends AppRoute | AppRouter,
-  TQuery,
-  TBody,
-  TParams,
-  TResponse,
-  TStatus = 200
->(
-  handler: Handler<TQuery, TBody, TParams, TResponse>
-): (all: RequestType2<TRoute, TQuery, TBody, TParams>) => Promise<{
-  status: TStatus;
-  body: { message: string; data: TResponse };
-}> {
-  return async (all) => {
-    const req: MonkeyTypes.Request2<TQuery, TBody, TParams> = {
-      body: all.body as TBody,
-      query: all.query as TQuery,
-      params: all.params as TParams,
-      raw: all.req,
-      ctx: all.req["ctx"],
-    };
-
-    const result = await handler(req);
-    const response = {
-      status: 200 as TStatus,
-      body: {
-        message: result.message,
-        data: result.data as TResponse,
-      },
-    };
-
-    return response;
-  };
-}
-
-type WithBody<T> = {
-  body: T;
-};
-type WithQuery<T> = {
-  query: T;
-};
-
-type WithParams<T> = {
-  params: T;
-};
-
-type WithoutBody = {
-  body?: never;
-};
-type WithoutQuery = {
-  query?: never;
-};
-type WithoutParams = {
-  params?: never;
-};
-
-type Handler<TQuery, TBody, TParams, TResponse> = (
-  req: MonkeyTypes.Request2<TQuery, TBody, TParams>
-) => Promise<MonkeyResponse2<TResponse>>;
-
-type RequestType2<
-  TRoute extends AppRoute | AppRouter,
-  TQuery,
-  TBody,
-  TParams
-> = {
-  req: TsRestRequest<TRoute>;
-} & (TQuery extends undefined ? WithoutQuery : WithQuery<TQuery>) &
-  (TBody extends undefined ? WithoutBody : WithBody<TBody>) &
-  (TParams extends undefined ? WithoutParams : WithParams<TParams>);
